+++ packages/server/src/services/db/models/Users.ts
@@ -1,4 +1,5 @@
-import { Column, Columns, Model } from "../helpers/orm";
+import { Column, Columns, Model } from "../orm";
+import { pool } from "../pool";
 
 export interface IUser {
   id: string;
@@ -10,13 +11,14 @@ export interface IUser {
 }
 
 export const Users = new Model<IUser>({
+  pool,
   schema: "public",
   table: "users",
   columns: new Columns([
     new Column({ name: "id", datatype: "uuid" }),
     new Column({ name: "fullname", datatype: "text" }),
     new Column({ name: "email", datatype: "text" }),
-    new Column({ name: "hashed_password", datatype: "text" }),
+    new Column({ name: "hashed_password", datatype: "text", hidden: true }),
     new Column({ name: "created_on", datatype: "timestamptz" }),
     new Column({ name: "updated_on", datatype: "timestamptz" }),
   ]),
