+++ packages/server/src/services/db/models/index.ts
@@ -1,0 +1,2 @@
+export { Users } from "./Users";
+export type { IUser } from "./Users";
