+++ packages/server/src/services/db/orm/Model.ts
@@ -1,27 +1,29 @@
-import { type QueryResultRow, type QueryResult } from "pg";
-import { getDbClient } from "../../db";
 import format from "pg-format";
-import { type Columns } from "./Columns";
+import type { Pool, QueryResult, QueryResultRow } from "pg";
+import type { Columns } from "./Columns";
 
 interface IModelConstructorArgs {
+  pool: Pool;
   schema: string;
   table: string;
   columns: Columns;
 }
 
 export class Model<TModelDef extends QueryResultRow> {
-  table: string;
+  pool: Pool;
   schema: string;
+  table: string;
   columns: Columns;
 
-  constructor({ schema, table, columns }: IModelConstructorArgs) {
+  constructor({ pool, schema, table, columns }: IModelConstructorArgs) {
+    this.pool = pool;
     this.schema = schema;
     this.table = table;
     this.columns = columns;
   }
 
   async findOne(): Promise<TModelDef> {
-    const client = await getDbClient();
+    const client = await this.pool.connect();
     const sql = format(
       /* sql */ `
         select %I from %I.%I limit 1
