+++ packages/server/src/services/db/pool.ts
@@ -1,0 +1,5 @@
+import { Pool } from "pg";
+
+export const pool = new Pool({
+  connectionString: process.env.DB_URL,
+});
